7699 move ScopedFilterset
